+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # ActFramework Change Log
 
 **1.8.20**
+* envMatches in ControllerByteCodeScanner shall be reset upon scanning new class #1065
+* Act-Test - support post XML encoded body #1062
 * API book - Module name shall include enclosing class #1059
 * API doc - line breaks in param description get removed #1060
 * metric command stop working #1058